+++ packages/lib/test/parent/parent.test.ts
@@ -1,6 +1,7 @@
 import { remove, set } from "mobx"
 import {
   detach,
+  findChildren,
   findParent,
   findParentPath,
   fromSnapshot,
@@ -113,6 +114,20 @@ test("parent", () => {
   expect(Array.from(getChildrenObjects(p.arr).values())).toEqual(p.arr)
   expect(Array.from(getChildrenObjects(p.p2!).values())).toEqual([])
 
+  expect(Array.from(findChildren(p, () => true).values())).toEqual([
+    p.arr,
+    p.arr[0],
+    p.arr[1],
+    p.arr[2],
+    p.p2,
+  ])
+  expect(Array.from(findChildren(p, (node) => node instanceof P2).values())).toEqual([
+    p.arr[0],
+    p.arr[1],
+    p.arr[2],
+    p.p2,
+  ])
+
   expect(getParentToChildPath(p, p)).toEqual([])
   expect(getParentToChildPath(p, p.p2!)).toEqual(["p2"])
   expect(getParentToChildPath(p.p2!, p)).toEqual(undefined) // p is not a child of p.p2
@@ -129,6 +144,11 @@ test("parent", () => {
   expect("p2" in p.$).toBeFalsy()
   expect(getParentPath(p2)).toBeUndefined()
   expect(Array.from(getChildrenObjects(p).values())).toEqual([p.arr])
+  expect(Array.from(findChildren(p, (node) => node instanceof P2).values())).toEqual([
+    p.arr[0],
+    p.arr[1],
+    p.arr[2],
+  ])
 
   // readd prop
   runUnprotected(() => {
@@ -138,6 +158,12 @@ test("parent", () => {
   })
   expect(getParentPath(p2)).toEqual({ parent: p, path: "p2" })
   expect(Array.from(getChildrenObjects(p).values())).toEqual([p.arr, p.p2])
+  expect(Array.from(findChildren(p, (node) => node instanceof P2).values())).toEqual([
+    p.arr[0],
+    p.arr[1],
+    p.arr[2],
+    p.p2,
+  ])
 
   // reassign prop
   runUnprotected(() => {
@@ -145,6 +171,11 @@ test("parent", () => {
   })
   expect(getParentPath(p2)).toBeUndefined()
   expect(Array.from(getChildrenObjects(p).values())).toEqual([p.arr])
+  expect(Array.from(findChildren(p, (node) => node instanceof P2).values())).toEqual([
+    p.arr[0],
+    p.arr[1],
+    p.arr[2],
+  ])
 
   // readd prop
   runUnprotected(() => {
@@ -152,6 +183,12 @@ test("parent", () => {
   })
   expect(getParentPath(p2)).toEqual({ parent: p, path: "p2" })
   expect(Array.from(getChildrenObjects(p).values())).toEqual([p.arr, p.p2])
+  expect(Array.from(findChildren(p, (node) => node instanceof P2).values())).toEqual([
+    p.arr[0],
+    p.arr[1],
+    p.arr[2],
+    p.p2,
+  ])
 
   // detach
   runUnprotected(() => {
@@ -160,6 +197,11 @@ test("parent", () => {
   expect(getParentPath(p2)).toBeUndefined()
   expect(p.p2).toBeUndefined()
   expect(Array.from(getChildrenObjects(p).values())).toEqual([p.arr])
+  expect(Array.from(findChildren(p, (node) => node instanceof P2).values())).toEqual([
+    p.arr[0],
+    p.arr[1],
+    p.arr[2],
+  ])
 
   // readd prop
   runUnprotected(() => {
@@ -169,6 +211,12 @@ test("parent", () => {
   })
   expect(getParentPath(p2)).toEqual({ parent: p, path: "p2" })
   expect(Array.from(getChildrenObjects(p).values())).toEqual([p.arr, p.p2])
+  expect(Array.from(findChildren(p, (node) => node instanceof P2).values())).toEqual([
+    p.arr[0],
+    p.arr[1],
+    p.arr[2],
+    p.p2,
+  ])
 
   // detach once more
   runUnprotected(() => {
@@ -195,6 +243,11 @@ test("parent", () => {
   expect(p.arr.length).toBe(3)
   expect(getParentPath(popped)).toBeDefined()
   expect(Array.from(getChildrenObjects(p).values())).toEqual([p.arr])
+  expect(Array.from(findChildren(p, (node) => node instanceof P2).values())).toEqual([
+    p.arr[0],
+    p.arr[1],
+    p.arr[2],
+  ])
   for (let i = 0; i < p.arr.length; i++) {
     expect(getParentPath(p.arr[i])!.path).toBe(i)
   }
@@ -240,6 +293,10 @@ test("parent", () => {
   expect(getParentPath(p2arr[1])).toBeDefined()
   expect(getParentPath(p2arr[2])).toBeUndefined()
   expect(Array.from(getChildrenObjects(p).values())).toEqual([p.arr])
+  expect(Array.from(findChildren(p, (node) => node instanceof P2).values())).toEqual([
+    p.arr[0],
+    p.arr[1],
+  ])
 
   runUnprotected(() => {
     p.arr[2] = p2arr[2]
@@ -248,6 +305,11 @@ test("parent", () => {
   expect(getParentPath(p2arr[1])).toBeDefined()
   expect(getParentPath(p2arr[2])).toBeDefined()
   expect(Array.from(getChildrenObjects(p).values())).toEqual([p.arr])
+  expect(Array.from(findChildren(p, (node) => node instanceof P2).values())).toEqual([
+    p.arr[0],
+    p.arr[1],
+    p.arr[2],
+  ])
 
   // detach
   runUnprotected(() => {
@@ -258,6 +320,10 @@ test("parent", () => {
   expect(getParentPath(p2arr[1])).toBeUndefined()
   expect(getParentPath(p2arr[2])).toBeDefined()
   expect(Array.from(getChildrenObjects(p).values())).toEqual([p.arr])
+  expect(Array.from(findChildren(p, (node) => node instanceof P2).values())).toEqual([
+    p.arr[0],
+    p.arr[1],
+  ])
 
   // set length
   runUnprotected(() => {
@@ -267,6 +333,7 @@ test("parent", () => {
   expect(getParentPath(p2arr[1])).toBeUndefined()
   expect(getParentPath(p2arr[2])).toBeUndefined()
   expect(Array.from(getChildrenObjects(p).values())).toEqual([p.arr])
+  expect(Array.from(findChildren(p, (node) => node instanceof P2).values())).toEqual([p.arr[0]])
 
   // adding to the array something that is already attached should fail
   const oldArrayLength = p.arr.length